snap

+++ crates/libs/windows/src/Windows/Win32/shobjidl_core/mod.rs
@@ -18781,7 +18781,7 @@ impl ISuspensionDependencyManager_Vtbl {
 }
 #[cfg(feature = "winnt")]
 impl windows_core::RuntimeName for ISuspensionDependencyManager {}
-pub const ITSAT_DEFAULT_LPARAM: usize = 18446744073709551615;
+pub const ITSAT_DEFAULT_LPARAM: usize = 18446744073709551615u64 as usize;
 pub const ITSAT_DEFAULT_PRIORITY: u32 = 268435456;
 pub const ITSAT_MAX_PRIORITY: u32 = 2147483647;
 pub const ITSAT_MIN_PRIORITY: u32 = 0;
